Refactor DB config and add initial migration
Updated `docker-compose.yml` to change the environment variable for the connection string and the working directory for the `ef-migrations` service. Updated `01-create-databases.sql` to create a new `StileStreamWmsProducts` database. Removed `ProductEntity.cs` and `ProductsConfiguration.cs`. Updated `ProductsDbContext.cs` to include `ConfigureShadowProperties` method. Added `ProductConfiguration.cs` for configuring the `Product` entity. Added initial migration files `20241001203525_Initial.cs` and `20241001203525_Initial.Designer.cs` for creating `OutboxMessages` and `Products` tables. Added `ProductsDbContextModelSnapshot.cs` to reflect the current state of the `ProductsDbContext` model.
