+++ StileStreamWms/init/sqlserver/01-create-databases.sql
@@ -1,2 +1,4 @@
 CREATE DATABASE StileStreamWms;
 GO
+CReate DATABASE StileStreamWmsProducts;
+GO
